update hardware config
- use unstable packages
  this required renaming some attr

add some amd graphics related settings to test if I can get around the ollama bug
that prevents allocation of additional gtt memory. This is done so that the model
can be loaded into GPU assigned ram ( right now only 4gb that are statically assigned
as vram via bios are used). This seems to slow down running the model signifcantly.

fly-by:
- remove qmk ( the keyboards all run zmk )
- remove networking override. it's not required anymore